Remove everything from azure-common that wasn't truly shared by several libraries, leaving only the exception classes in azure-common, the rest is moved to azure-mgmt-common. The aim is to have azure-common as stable as possible with regards to backwards compatibility, and the stuff from azure-mgmt-* is subject to change at this time.
