+++ azure-mgmt-common/README.rst
@@ -1,0 +1,4 @@
+Microsoft Azure SDK for Python
+==============================
+
+This is the Microsoft Azure Resource Management common code.
