+++ azure-mgmt-common/azure/mgmt/common/arm.py
