+++ azure-mgmt-common/setup.cfg
@@ -1,0 +1,6 @@
+[bdist_wheel]
+universal=1
+
+[install]
+single-version-externally-managed=1
+record=RECORD.txt
