+++ azure-mgmt/tests/mgmt_testcase.py
@@ -27,6 +27,14 @@ from testutils.common_recordingtestcase import (
 import tests.mgmt_settings_fake as fake_settings
 
 
+should_log = os.getenv('SDK_TESTS_LOG', '0')
+if should_log.lower() == 'true' or should_log == '1':
+    import logging
+    logger = logging.getLogger('azure.mgmt.common.filters')
+    logger.setLevel(logging.DEBUG)
+    logger.addHandler(logging.StreamHandler())
+
+
 class HttpStatusCode(object):
     OK = 200
     Created = 201
